refactor: MVC 패턴으로 패키지 정리
