+++ src/main/java/racingcar/view/ResultView.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.view;
+
+import racingcar.domain.Cars;
 
 public class ResultView {
