+++ src/test/java/racingcar/domain/CarsGeneratorTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
