+++ src/test/java/racingcar/domain/CarsTest.java
@@ -1,8 +1,9 @@
-package racingcar;
+package racingcar.domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.strategy.AlwaysMoveStrategy;
 
 import java.util.List;
