+++ src/test/java/racingcar/domain/strategy/AlwaysMoveStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.strategy;
 
 public class AlwaysMoveStrategy implements MoveStrategy {
