+++ src/test/java/racingcar/domain/strategy/RandomMoveStrategyTest.java
@@ -1,9 +1,10 @@
-package racingcar;
+package racingcar.domain.strategy;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racingcar.domain.Car;
 
 import java.util.Random;
